Move Holiday themes into the Region section
The Holiday themes toggles now live as a section on the Region page
instead of a separate top-level Settings entry. Putting the holidays
next to language / temperature / distance reflects that the holidays
worth seeing are largely a function of where you live — same reason
the v3 plan (TODO in HolidayTheme.kt) will derive a country picker
from this section.

The 23 per-holiday switches keep their existing on-by-default + tactful
opt-out behaviour. The follow-up TODO spells out the next step: tag
every theme with a `countries: Set<String>`, gate visibility by the
user's chosen country set (defaulted from Region), and reframe the
section copy so opting out of Mother's Day / Father's Day /
Remembrance Day reads as the gentle gesture it should be.

- Drop the standalone SettingsRoute.Holidays page; HolidaysContent moves
  into RegionContent as a fourth SectionCard. The route saver already
  falls back to Root for unknown names, so old installs that had the
  Holidays page open at process death restore cleanly.
- Remove `settings_root_holidays{,_subtitle}` strings — they were
  English-only and the row they labelled is gone.
- Bump the settings_region snapshot viewport to 1600dp so the four
  cards + every holiday row fit in one frame; settings_holidays.png is
  no longer captured.
